pyproject.toml
[tool.poetry]
name = "rb2024-lookanotherdev"
version = "0.1.0"
description = "Minha submissão para a Rinha de Backend 2024"
authors = ["André Carvalho <alac1984@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "rb2024_lookanotherdev"}]

[tool.poetry.dependencies]
python = "^3.12"
starlette = "^0.37.1"
uvicorn = "^0.27.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
dataclasses-json = "^0.6.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
flake8 = "^7.0.0"
black = "^24.2.0"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
